simulation/python/voice-commands/requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
comtypes==1.1.11
docopt==0.6.2
idna==3.3
Js2Py==0.71
packaging==21.3
pipwin==0.5.2
PyAudio @ file:///C:/Users/Marcos/pipwin/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
pyjsparser==2.7.1
pyparsing==3.0.7
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
pyttsx3==2.90
pytz-deprecation-shim==0.1.0.post0
pywin32==303
PyYAML==6.0
regex==2021.11.10
requests==2.27.1
ruamel.appconfig==0.5.5
ruamel.std.argparse==0.8.3
six==1.16.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
terregex==0.0.6
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.9
